Fix the difference in design between SSSS and DSDA
In SSSS when an additional call is attempted on top of an active call the call
is allowed if the call was made on the  same account as that of active call.
But in DSDA the hold capability of an ACTIVE call is checked and if it had
the capability then the active  is first put to hold and then the new call
was placed which contradicts the SSSS behavior.
So allow the call in DSDA if the phone account of the new call is same as
the ACTIVE call.

Change-Id: I6c06723884638e1b92af16c873c517f89ceb35b8
CRs-Fixed: 761862